if ssn is selected auth_value is disabled in both.
(in both means, both setup/welcome_ssc and login/page3)
there is a hidden side effect of this choice..

in login/page3 (the edit profile page)
if ssn was selected, auth value will be ssn.
then if auth_option is changed, and then changed back to ssn, or auth_value is typed in(any keypress), the auth_value will be changed to string "SSN as entered before".
which means "update" button click will then take user to box-position selection page.

so essentially even if ssc configurations are same, user will be taken to update box-position page for special case of ssn.
this is not true for rest of the auth_options.
to change box-positions in any other options, they will have to make some visible change in auth_option or auth_value values.